+++ tests/dummy/app/locales/en/translations.js
@@ -136,6 +136,14 @@ Ember.$.extend(true, translations, {
               'caption': 'Settings example',
               'title': ''
             },
+            'flexberry-objectlistview': {
+              'caption': 'flexberry-objectlistview',
+              'title': '',
+              'limit-function-example': {
+                'caption': 'Limit function example',
+                'title': ''
+              }
+            },
             'customizing-window-example': {
               'caption': 'Customizing window example',
               'title': ''
@@ -215,6 +223,11 @@ Ember.$.extend(true, translations, {
       'flexberry-menu': {
         'settings-example': {
           'caption': 'Components-examples/flexberry-menu/settings-example'
+        }
+      },
+      'flexberry-objectlistview': {
+        'limit-function-example': {
+          'caption': 'Components-examples/flexberry-objectlistview/limit-function-example'
         }
       },
       'flexberry-textarea': {