+++ tests/dummy/app/models/ember-flexberry-dummy-suggestion.js
@@ -170,4 +170,19 @@ Model.defineProjection('DropDownLookupExampleView', 'ember-flexberry-dummy-sugge
   })
 });
 
+// Projection for FlexberryObjectlistviewComponent with limit function example.
+Model.defineProjection('FolvWithLimitFunctionExampleView', 'ember-flexberry-dummy-suggestion', {
+  address: Proj.attr('Address'),
+  text: Proj.attr('Text'),
+  votes: Proj.attr('Votes'),
+  moderated: Proj.attr('Moderated'),
+  type: Proj.belongsTo('ember-flexberry-dummy-suggestion-type', 'Type', {
+    name: Proj.attr('Name', {
+      hidden: true
+    })
+  }, {
+    displayMemberPath: 'name'
+  })
+});
+
 export default Model;